requirements.txt
click==8.1.3
easydict==1.9
matplotlib==3.5.1
MinkowskiEngine==0.5.4
numpy==1.20.3
open3d==0.13.0
pykeops==2.1
pytorch_lightning==1.5.6
PyYAML==6.0
scipy==1.7.3
torch==1.10.1+cu113
